xserver libX11 libxcb mesa xcb-proto glproto git update March 31 2017
xserver          commit e4d0757fc2620fd3182448c5e4b5a55a509594d3
libxcb           commit ee9dfc9a7658e7fe75d27483bb5ed1ba4d1e2c86
libxcb/xcb-proto commit 97fbba25baf241cbcda882739db423b1d36e440a
libX11           commit b856d5d929047d1ea169814d56e43784ea404c83
glproto          commit 500b137bbdba1cd8ae943849fef6e79fdd4eca40
mesa             commit 02112c3ef7716955c889b6d10fb569b028a33f0e

+++ mesalib/src/util/bitscan.h
@@ -52,7 +52,7 @@ extern "C" {
 #define ffs __builtin_ffs
 #elif defined(_MSC_VER) && (_M_IX86 || _M_ARM || _M_AMD64 || _M_IA64)
 static inline
-int ffs(unsigned i)
+int ffs(int i)
 {
    unsigned long index;
    if (_BitScanForward(&index, i))
@@ -62,14 +62,14 @@ int ffs(unsigned i)
 }
 #else
 extern
-int ffs(unsigned i);
+int ffs(int i);
 #endif
 
 #ifdef HAVE___BUILTIN_FFSLL
 #define ffsll __builtin_ffsll
 #elif defined(_MSC_VER) && (_M_AMD64 || _M_ARM || _M_IA64)
 static inline int
-ffsll(uint64_t i)
+ffsll(long long int i)
 {
    unsigned long index;
    if (_BitScanForward64(&index, i))
@@ -79,7 +79,7 @@ ffsll(uint64_t i)
 }
 #else
 extern int
-ffsll(uint64_t val);
+ffsll(long long int val);
 #endif
